Arreglando bugs en m_authadd:  * Ahora remueve los permisos anteriores del mismo grupo si se otorgan permisos nuevos  * Arreglado crasheo cuando se borraban privilegios inexistentes

+++ modules/m_authadd.php
@@ -15,11 +15,20 @@ class amodkey{
 		$core->registerCommand("listpriv", "authadd", "Lista los privilegios de un usuario. Sintaxis: listpriv <usuario>");
 		$core->registerCommand("addpriv", "authadd", "Da privilegios a un usuario. Sintaxis: addpriv <usuario> <privilegios> <sector>",9);
 		$core->registerCommand("delpriv", "authadd", "Quita privilegios a un usuario. Sintaxis: delpriv <usuario> <privilegios> <sector>",9);
+		$core->registerCommand("listusers", "authadd", "Lista los usuarios actualmente registrados");
 
 	}
 	
+	public function listusers(&$irc, &$data, &$core){
+		$users = ORM::for_table('users')->find_many();
+		$r="Usuarios actualmente registrados: ";
+		foreach($users as $val){$r.="{$val->username}, ";}
+		$r = trim($r,", ");
+		$irc->message(SMARTIRC_TYPE_CHANNEL, $data->channel, $r);
+		
+	}
 	public function register(&$irc, &$data, &$core){
-		if(isset($data->messageex[2])){
+		if(!isset($data->messageex[2])){
 			$irc->message(SMARTIRC_TYPE_QUERY, $data->nick, "Faltan parámetros. Sintaxis: register <usuario> <contraseña>");
 		}
 		$user = ORM::for_table('users')->create();
@@ -32,6 +41,7 @@ class amodkey{
 		$priv->rng = "0";
 		$priv->sec = "*";
 		$priv->save();
+		$irc->message(SMARTIRC_TYPE_QUERY, $data->nick, "Registrado exitosamente.");
 	}
 	
 	public function listpriv(&$irc, &$data, &$core){
@@ -50,6 +60,8 @@ class amodkey{
 		if(!isset($data->messageex[3])){$irc->message(SMARTIRC_TYPE_CHANNEL, $data->channel, "\00305Error:\003 Faltan parámetros!!");}// Por que enviar un mensaje de error al usuario es mucho trabajo
 		if($data->messageex[2]>9){$irc->message(SMARTIRC_TYPE_CHANNEL, $data->channel, "\00305Error:\003 Error de usuario. Inserte otro usuario y presione enter. (No se pueden otorgar privilegios de nivel 10!!)");}
 		$user = ORM::for_table('users')->where('username', $data->messageex[1])->find_one(); 
+		$k = ORM::for_table('userpriv')->where('uid', $user->id)->where('sec',$data->messageex[3])->find_one();
+		if(method_exists($k, "delete")){$k->delete();}
 		$priv = ORM::for_table('userpriv')->create();
 		$priv->uid = $user->id;
 		$priv->rng = $data->messageex[2];
@@ -64,7 +76,7 @@ class amodkey{
 		$user = ORM::for_table('users')->where('username', $data->messageex[1])->find_one();
 		if(!isset($user->id)){return 0; } //el usuario no existeer 
 		$userpriv = ORM::for_table('userpriv')->where('uid', $user->id)->where("rng", $data->messageex[2])->where("sec", $data->messageex[3])->find_one();
-		$userpriv->delete();
+		if(method_exists($userpriv, "delete")){$userpriv->delete();}
 		
 		$irc->message(SMARTIRC_TYPE_CHANNEL, $data->channel, "Se han otorgado los privilegios.");
 	}
